+++ switch_interface/detection.py
