+++ switch_interface/resources/layouts/alphabet_symbols.json
