+++ switch_interface/resources/layouts/basic_test.json
