+++ switch_interface/resources/layouts/media_controls.json
