+++ switch_interface/resources/layouts/numeric_pad.json
