+++ tests/test_layout_loader.py
@@ -1,4 +1,4 @@
-from myproject.kb_layout_io import load_keyboard
+from switch_interface.kb_layout_io import load_keyboard
 
 def test_default_load():
     kb = load_keyboard()
